requirements.txt
# Core scientific
numpy
scipy==1.14.0
pandas==2.2.2
matplotlib==3.9.1
seaborn==0.13.2
scikit-learn==1.5.1
pyarrow==17.0
joblib==1.4.2
PyYAML==6.0.1
pydantic==2.8.2
six==1.16.0

# Deep learning
torch==2.4.0 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.19.0 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu121
torchmetrics==1.4.0
torchdiffeq==0.2.4
torch_geometric==2.5.3

# CUDA-specific
# must be installed with -f or --extra-index-url
dgl -f https://data.dgl.ai/wheels/torch-2.4/cu121/repo.html
# pytorch3d 
git+https://github.com/facebookresearch/pytorch3d.git@V0.7.8

# Other ML tools
xgboost==2.1
optuna==3.6.1
wandb==0.17.5
tensorboard==2.18.0
nibabel==5.2.1
nitransforms==23.0.1
pip install torchdata==0.7.1
pip install  dgl -f https://data.dgl.ai/wheels/torch-2.4/cu121/repo.html
# conda install -c dglteam/label/th24_cu121 dgl
pytest
scikit-image==0.25.2
umap-learn==0.5.9
geomstats==2.8.0
pyvista==0.46.3
statsmodels==0.14.5